Add files via upload
Assembly Manuals for the Barb-A-Watt QRP Power/SWR Meter and Dummy Load. There are two construction and assembly manuals. 
Version 2.3 is the latest manual for the Barb-A-Watt Rev 1.2 Board (the second design) which uses a SMD PIC16F1938 processor and a 9v Battery.
Version 3.01 is the lates manual for the latest Version 2.0 Main Board Barb-A-Watt which uses a 28Pin DIP PIC controller and operates with 2 x AAA batteries.
The operation and function of both Barb-A-Watt versions are exactly the same.